requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
langchain==0.3.7
langchain-core==0.3.15
langchain-text-splitters==0.3.2
langchain-community==0.3.5
langchain-groq==0.2.1

